Make union types more efficient by eliminating unnecessary redundancy.
This change modifies union types to always leave structural subtypes in
the union.  This can result in additional types errors if the previously
collapsed type provides a definition for a property or if the collapsed
type was @struct (as any @struct member of the union causes the member
to be considered @struct).

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=242546590