Add a new gwt entry point to help slice up JS Doc for ES5 -> ES6 class conversions.
Parameters need to move to the constructor and everything else stay on
the class (e.g. @implements, @template, description, etc).

The compiler is the only thing that understands how to parse JS Doc
correctly and it is very hard to reimplement this logic. I think this is
the best of some not fantastic options.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=242717398