Defer declaring goog.requireType imports
Unlike a typical forward reference to a type, which are resolved through NamedType, references to goog.requireType types in goog.modules are currently broken. This is because in cases like

  const Data = goog.requireType('mod.Data');
  let /** !Data */ data;

TypedScopeCreator has already traversed the name "Data" local to the module. Making a NamedType to resolve "Data" is ineffective - we never declare the name "Data" in the module scope.

Two ways of fixing this problem are:
  1) (this CL) visit all goog.requireType imports after all module scopes have been created, but before type resolution.
  2) modify type resolution to make "Data"'s NamedType (and types referencing properties of "Data") resolve in the scope of 'mod.Data', not the current module. This seems more complicated. JSTypeRegistry would have to know that the type name "Data" should actually be resolved in some other scope completely when inside a given module.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=242949071