Migrate standardized History API to html5.js from ie_dom.js
IE10 and IE11 also support passing a string parameter as a parameter
to History.prototype.go function and passing an optional delta parameter
to History.prototype.back both of which are easily replaced with standards
equivalents (i.e. using "window.location.href = ..." or "history.go(delta)").

The externs for these non-standard parameters has been removed as part
of this PR as it seems very unlikely that there is limited if any use
of these parameters any more.