Follow aliases and module imports for @define initializers
The criterion for a valid @define initializer is that it be a statically-computable expression containing only literals or other @defines, but existing code used a naïve string-keyed map to determine whether an RHS name was a @define, which fails to understand modules and aliases.

To fix this, we now do an extra pass over all global declarations to find any additional valid aliases, repeating until a fixed point is found.

Also moves isValidDefineValue out of NodeUtil, since this is the only place it should ever be used.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=243703850