Move the Location externs from ie_dom.js to html5.js
The Location interface has been standardized and can be moved to
the standards based externs. This means that google/elemental2 will
no longer need to to have local patched to include externs for this type.

The @see references have been updated to point at the new spec and the
documentation has been copied from the Elemental2 externs patch (which
was ultimately sourced from the spec).

The one type change is that `function Location()` was previously
annotated with @constructor but is now annotated with @interface to both
match the spec and generate better code in google/elemental2. Given that
this is not an object that a end-user can instantiate it is not expected
to have any impact.