Make JsFileParser more resilient on files with multiple goog.provide/goog.module strings.
This patch makes first module type that's recognized by JsFileParser as the module type and also stops recording the provided namespace in cases of module type conflict.

As a result of module type conflicts that are due to inaccurate parsing done via JsFileParser are less dangerous since the top one will win which is most likely the correct one.

This fixes most of the issues in compiled IntegrationTest that is cross compiled with J2CL.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=244714967