Adds typing support for `this` and `super` in static methods.
The main changes are in `TypedScope` (to generate the type for `super`) and `TypedScopeCreator` (to determine the type of `this` for static methods).

This changes applies to all functions declared as properties on constructors and interfaces. That includes ES5 constructors. However, this does not mean any kind of support for class side inheritance on ES5 constructors.

The majority of the changes are test updates to accommodate the new type information.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=245573045