Make TypedScope never return qualified names whose root is shadowed.
For example, given the following:
GLOBAL SCOPE: {a, a.b} and LOCAL SCOPE: {a}

Looking up "a.b" in LOCAL SCOPE used to return the global "a.b". Now it returns null.

This includes some fixes in TypedScopeCreator to declare 'this' properties and object literal keys on the correct scope.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=248767080