Move standardized document methods to w3c_dom1.js and remove useless overrides of same methods by HTMLDocument
The open(...), close(), write(...) and writeln(...) methods on Document are
now part of the standard and have been moved to w3c_dom1.js to reflect this.

The typing of Document.prototype.close has been aligned with the specification
and with the overriding close methods in HTMLDocument.

The overrides  of close(), write() and writeln() in HTMLDocument have been
removed. This is done due to limitations of the jsinterop-generator tool that
is more fully described in google/jsinterop-generator#35 While the specifications
explicitly indicate that these overrides are present and annotates (requiring the
[OverrideBuiltins] attribute be applied to the containing interface in
WebIDL) it is not expected that this has any practical impact on programs
type checking against the externs.

Closes https://github.com/google/closure-compiler/pull/3399.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=251955300