Automated g4 rollback of changelist 251282822.
*** Reason for rollback ***

Breaks sourcemap generation in JS version of the compiler

*** Original change description ***

Switch the JS implementation of the transpiler to be created using J2CL rather than GWT.

This change restructures the generation process a bit because we can now optimize using the compiler itself and we don't have to unzip GWT JARs.

There's a 20% size increase in the transpiler following this change. As a next step this could be mitigated by better tailoring the resource files shipped with the transpiler, most of which are unused.

***

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=252623480