Allow aliasing messages onto objects
This is an extension of the earlier work to support exporting
messages. The changes here are done to account for when declareLegacyNamespace
is being used and the code has already been rewritten.

That is, the code was `exports = {MSG_FOO};` but after rewriting is
`a.b.SomeClass.MSG_FOO = module$contents__MSG_FOO`.

There isn't a great way to know this rewriting has happened though, so
technically this change allows any local variable to be aliased onto an object.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=252694073