In `DisambiguateProperties`, stop conflating ctor types with the interface types they declared their instance types to implement.
We want conflation between instance types and their implemented interfaces, but ctors types don't have any connection to those interfaces. The mistake was that when getting the ctor type of a ctor type, which is where interfaces would be stored, we were returning the input ctor type; the ctor of a ctor is not itself.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=253056130