Attach JS source comments to the AST nodes
With this change, we give fixjs (and internally the linter) a JS compiler option
which can be used to preserve comments present in the JS source code. The
comments associated with a node gets added to the node's Prop linked list.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=254830379