Fix for incorrect compiler warning  about missing inline JSDoc info
The problem was that for every default param, we have a node DEFAULT_VALUE
which holds the lhs(identifier NAME) and rhs(default value). We attach the
inline JsDocInfo to the lhs (NAME node), but  we push its parent DEFAULT_VALUE
into the function param list. So while searching for JsDoc info in the function's param list, we do not look into the child NAME node which holds the JsDoc info.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=255040368