Avoid converting $.foo to $['foo'] for method calls
Polymer classes have a `$` property that is used as a dictionary-like object. Client code often does stuff like `this.$.someProp` to access these properties. Left unmodified, JSCompiler would complain about these, because it doesn't have sufficient type information on `this.$` to know what properties exist on it. So, the Polymer pass converts `this.$.someProp` to `this.$['someProp']`, which JSCompiler won't complain about, because it doesn't look into the values used for bracket access.

The logic of the Polymer pass sometimes does this conversion when it shouldn't, because it cannot tell which properties named `$` really belong to a Polymer object and which don't. This is a problem for some method calls that are accessed as something.$.someMethod().

This CL makes the Polymer pass a bit smarter by avoiding the conversion when the `something.$.someProp` is the callee of a function call. None of the properties of a Polymer `$` property should be functions to call, so this avoids incorrect conversion for those cases.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=255083352