Introduce an options for `--assume_getters_and_setters_are_side_effect_free`.
The assumption is implemented by suppressing getter/setter property collection

The default value is initially `true` as that's the current assumption of the compiler, and we don't have code to operate any differently. However, once such code is added and vetted the default will be changed to `false`.

This option is not intended to be set by most users.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=255697780