Improve .i.js generation for Polymer Behaviors.
JS Compiler incremental compilation works by simplifying down each file to just what's needed for type-checking, so that it runs faster.

By default, object literal property right-hand-sides are simplified to "0", since the value doesn't usually matter. However, in the case of Polymer Behaviors the RHS does matter, because the PolymerPass uses it to understand which properties the behavior is going to set on the Polymer Elements that apply it.

This change preserves any property called "properties" on an object literal that is annotated with @polymerBehavior, and simplifies it so that only the necessary type information is preserved.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=255987942