Consider side-effects of getters and setters when tracking property read/writes in `RemoveUnusedCode`.
Previously we could assume that if a property was never read it never needed to be written to. The possibility of a setter being assigned to undermines this assumption.

Additionally, assumptions were being made about when parts of expressions were implicitly safe to remove. The possibility that getters could have side-effects changes this.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=258450431