Change the `tracerOutput` option to be a `Path` rather than a `PrintStream`.
Streams need to be lifecycle managed which is difficult/impossible for one being set on the options object. Using a `Path` allows us to create the stream as needed and safely close it in the same context.

Additionally, the options should ideally be serializable (in the sense of a config file). The isn't possible for a `PrintStream` but is trivial for a `Path`.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=258451272