Create and apply new tokens ITER_SPREAD, OBJECT_SPREAD, ITER_REST, and OBJECT_REST to differentiate/replace SPREAD and REST.
These tokens are not currently used but are put in place throughout the compiler so that when they appear in the AST behaviour will not change. The update to the parser to generate nodes using these tokens will follow.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=259644219