Fix a couple of types in the Promise polyfill implementation.
`resolve` and `reject` are defined as receiving an optional parameter, but the polyfill uses functions with a mandatory parameter. These fail to warn due to a bug around type inference for object literal parameter types. This is a cleanup to unblock fixing this bug.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=261004826