Allow duplicate symbol declarations in the JSModuleGraph
Currently when building the map of <Symbol, CompilerInput> we don't check if
there's a pre-existing entry. This has the effect of silently dropping all but
the last file in the list. While having multiple inputs declare the symbols is
invalid, it's shouldn't be silently ignored. By allowing them through the
JSModuleGraph we can accurately check for dupes later on in the compilation.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=264408309