Automated g4 rollback
Original change description:

Fix bug where @suppress on a property write would be treated as the canonical prop JSDoc

Now JSDoc without any annotations that are meaningful to the type system (e.g. @const, @type, @private) is not attached to properties in the type system, as its unclear whether or not that JSDoc is meant to 'declare' the property.

***

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=264975482