Allow goog.forwardDeclare in externs when natively checking modules
We still require a definition of 'goog' in the JS code, but back off the usual warning for referencing a non-extern in the externs and the synthetic extern definition of 'goog'. This is because defining goog in the externs causes optimization back-off.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=269875820