Remove incorrect error reporting of unused template variable declarations.
The usage of a template variable does not always occur in the JsDoc block where the template is defined, e.g.

  class A {
    /** @template T*/
    method(/** T */ param) {
    }
  }

or

  class A {
    /** @template T*/
    method() { .... /** @type {T} */ (expr) ....
    }
  }

but the check is only done within the confines of a JsDoc annotation comment and it errors even when the template is actually used.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=272037489