Use property-absence tests to allow loose property access
The compiler considers `if (!goog.isDefAndNotNull(x.y))` as a property _presence_ test for `x.y` because it only looks at the `goog.isDefAndNotNull` call, and not what is being done to the result.  It then backs off from strict missing property warnings.  When this is replaced with `if (x.y == null)` the compiler considers it a property _absence_ test and does _not_ back off.  There's already a function in `NodeUtil` to check for property absence, but it's been unused since NTI was deleted.  This change now calls that function right next to the presence check and either will now allow backoff.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=272995616