Fix missing super call fixer
The fixer was trying to use the error node itself as a function, but the node the error was at was a member function definition. The first child of that node is the function, which can be used to extract the body.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=276774278