Prevent inlining function declarations in ClosureRewriteModule.
This causes problems when the function is used before the declaration, but the declaration is rewritten to an expression, which no longer hoists.

There are two possibilities for handling this: (1) hoist the declaration before inlining, or (2) don't inline function declarations.  The inlining should no longer be necessary since AggressiveInlineAliases should be able to handle it.  The alternative of manually hoisting in ClosureRewriteModules requires being extra careful about inserting the hoisted node after any potential redefinition of the `exports` object (though it's possible that this would prevent inlining in the first place).

There's a significant amount of fallout from this change.  The assumption that there would not be an alias for exported functions has been baked into a number of places, such as Dossier, which needs to record the extracted names as exported by the module, so that the exported symbol can be traced back to the real symbol's jsdoc.  Clutz's output also grows by quite a bit.

This is a clean roll-forward of 46937db now that internal tests are fixed.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=277175011