Memoize results for some tree inspection to speed up inlining
Before this change, callMeetsBlockInliningRequirements used
NodeUtil.has(...) to check for the absence of "eval" and function
definitions at the call site, which are both vanishingly rare. Since
predicates that return false cause worst-case performance for
NodeUtil.has(...), this used a large percentage of the processing
time.

This change caches the results of finding "equals" nodes and which
nodes contain which functions. The caches live for the lifetime of the
FunctionInjector, so new FunctionInjectors should be created after the
AST has been manipulated.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=277473647