Enable the check against using bounded generics before they're ready.
This diagnostic is to keep people from using the new feature while it's under development.

There was a logic error in the `WarningsGuard`s that effectively disabled the check for all users. This change simplfies the implementation to make the implemnetation less error-prone as well as correct.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=278426460