Roll back the restrictByNot* methods on ProxyObjectType
For TemplatizedType, the reference is always non-nullable, so these do nothing.  For TemplateType, if the bound is nullable then this could make a difference, but it's a toss-up which option is better (probably leaving it un-narrowed).  For NamedType, this will only be more correct if the type is actually resolved.  But then the behavior is inconsistent, which is also not great.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=278976837