Don't crash in NamedType.resolveInternal
NamedType.resolveInternal was implicitly depending on toMaybeObjectType()
always returning non-null when isObjectType() was true.

This is not always true (such as for unions where each union member is an
object). Changing the type system to make this true may be a reasonable change
in the future, but in the meantime, we shouldn't crash here.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=281401837