Prevent generating keywords as identifier names in the DefaultNameGenerator
We fail to detect the keywords 'yield', 'let' and 'await' in the scanner in
TokenStream.java. This causes them to be not reserved inside the name generator,
and we end up generate identifiers with those names.

Changing the TokenStream to be aware of language mode might be runtime-expensive as
Scanning is performance sensitive.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=282382147