Narrow * and ? to Object if goog.isObject returns true, rather than NoObjectType
For * this is clearly the right thing to do, since * is a very strict type, and narrowing to (essentially) bottom is too permissive.  For ? it's a little less clear (CheckedUnknown would not be an unreasonable option), but there doesn't seem to be anything depending on the looser narrowing, so we should be able to land this as is.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=282776614