Fixes compiler crash when using for-of without externs
Previous to this, the compiler would crash when for-of was used when no externs were provided:

```
for (let x of []) {}
```

This happened because the iterator `next` type was set to null (it's usually a function), then the return type was attempted to be accessed, which crashed the compiler. This change checks for nulls before access, and uses `?` rather than null to represent the missing types.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=286231758