Switch an assertion about a type being resolved into a call to resolve the type.
In a small number of edge cases this check was crashing the compiler. There is an effort to ensure that all types are always resolved during this stage of the compilation, but there's short-term value in getting this working again this way.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=286484992