rule-based: Ask IMF to reset the engine when using rule-based.
Right now, the extension may need to reset the state of an engine w/o
deactivating it. For example, if the user starts voice input, we need
to reset the engine state (voice input is not an engine; if it was an
engine, everything will work because we would deactivate the current
engine before starting voice input).

So far, this has been fine because the extension contained both UI
and IME code. But when we move to a native model, UI will no longer
contain IME code, so it will need to signal to IMF to reset the state
of the engine.

We added a new private API in [1] which we can use.

[1] https://chromium-review.googlesource.com/c/chromium/src/+/1971047

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=287914395