Sync compiler FeatureSet with AST contents.
Previously the compiler's FeatureSet was always all possible features indicated by
`--language_in`, minus any that had been compiled away so far, even if that included features that never appeared anywhere in the input externs and sources.

This made it difficult to update passes for individual features independently,
because the PhaseOptimizer would report an error or skip passes for features included
in `--language_in` even when they weren't anywhere in the input.
To add tests for a new feature, you'd have to define a custom language mode that included
that feature but no other not-yet-fully-supported features.

CHANGED: added computed properties feature in ClosureOptimizePrimitives.java to compiler feature set when it is added to the node.

***

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=288795497