Add type inference and type check tests for nullish coalesce
This was done by copying from tests for logical or.
This is to see what types of behaviour we should be checking for.
The tests have been modified to pass for now but the implementation for type inference is not yet in place. We will update the tests and add more specific test cases with implementation.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=289548173