No-op: Convert `TypeMismatch` into an AutoValue.
This CL cleans up the API for using, creating, and collecting `TypeMismatch` instances. The largest part of this is storing a `Node` for each mismatch rather than a `JSError`. Doing so eliminates a bunch of lazy evaluation code that was never actually used. No `JSError` is needed because mismatches are primarily for internal consumption.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=292619102