In the graph-based disambiguator, add type mismatches as edges in the type graph rather than marking those types as invalidating.
This change makes is possible to disambiguate properties on types involved in mismatches. As a result, properties of structural types can now be disambiguated, where that was effectively impossible before, since all structural matches are recorded as mismatches. It works because a mismatch doesn't necessarily mean a type's properties are ambiguous, just that they must be conflated with the part of the type lattice on the other side of the mismatch.

This approach is much more aggressive than the representative-based disambiguator.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=294518693