Add transpiler_mode flag to JSCompilerRunner.
The TL;DR; of the rational is that we would like to combine JS Transpiler into JSCompiler both to cut down on the number of transpilers and to reuse the JSCompiler worker during compilation.

To do this, the cl adds a transpiler_mode flag to JSCompiler that is responsible for configuring JSCompiler options for transpilation.

- Adds transpiler_mode doc_level=secret flag.
- Adds support for writing an inline source map to source file output in JSCompilerRunner

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=296046992