Use "exports" as the original name when rewriting "exports = ... " assignments
Currently when ClosureRewriteModule rewrites a directly "exports = ..."
assignment (note: _not_ a property on exports), it uses the full module
namespace as the original name. This can be problematic if tooling needs to
understand what the original code was since it's not representative of what it
was.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=299216362