obey @suppress in @fileoverview on JSDoc errors
JSDoc errors are attached to Nodes that don't have the script
as an ancestor.  When trying to find the relevant @suppress
and we haven't found a script ancestor, do an additional check
based on filename to search for any relevant @suppress.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=301856918