Add a new 'allow_extending_value' field to the conformance proto.
Some custom conformance checks use the 'value' field as a whitelist of
types, so it makes sense for subconfigs to add additional types to the
whitelist.  One example of this is closure:unknownProps, which is
currently only in Closure Library's private conformance file, and is
copy-pasted in many other conformance configs as a result, due to (1)
not being globally clean, and (2) needing to add additional whitelisted
types (specified in the 'value' field).  With this option, we can
finally expose that check in the base config with an 'only_apply_to'
restriction to Closure Library.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=302738897