Implement a pass to check for references to fully qualified namespaces.
This pass implements a more precise version of the "missing" check in
CheckMissingAndExtraRequires, by leveraging information about namespaces
provided in the full TC.

This check only affects modules. It flags all code and type references
to fully qualified names for which the corresponding namespace is
provided in the TC. It won't flag those that do refer to a namespace
when the providing file is absent from the TC (e.g. for a library with
incomplete dependencies). Such references are already flagged in
VarCheck (for code) and TypeCheck (for types).

False positives and negatives may be produced when `@template`
annotations are used, due to difficulties in identifying references to
template parameter names. We don't believe these to be prevalent enough
to merit additional effort at this time.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=302779919