Support fixing "JSC_JSDOC_REFERENCE_TO_FULLY_QUALIFIED_IMPORT_NAME" in `ErrorToFixMapper`.
This change consolidates the logic for various diagnostics between those reporting about JSDoc and those reporting about executable code. Doing so is required to get the right behaviour in the fixer.

The CL also switches to using the compiler to retrieve the SCRIPT node containing the diagnostic to fix. JSDoc ASTs are not connected to the SCRIPT that contains them, so the have to be located this way.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=305531275