Attach trailing inline comments to the correct argument or formal parameter
Currently, JSCompiler attaches the trailing comments present after an argument or formal parameter to the next argument or formal parameter after it. This changes the semantic meaning of the comments as they get attached to the wrong nodes, and also causes problems when there are trailing comments on the last argument or formal parameter.

This CL fixes that. It also add a boolean data member in NonJSDocComments to indicate whether any attached comment is trailing comment to the attached node or not. This boolean will be used by the code printer to preserve the location of the comment while emitting a node.

PiperOrigin-RevId: 308206580