Ensure all `SuggestedFixes` that depend on additional requires contain a copy of the new require.
Previously, only the first fix in a file that depended on a new require (A) would include it; the others assumed that (A) would be applied and define the name for them. In situations where the (A) conflicted with another fix, the require wasn't added, leaving the other fixes in a bad state.

To support many duplicated `CodeReplacements` for requires, identical `CodeReplacements` are no longer considered in conflict. They will just be deduplicated when being applied.

PiperOrigin-RevId: 312122849