AstValidator: optional chains cannot start with `super`
Actually AstValidator was already reporting an error as it should,
thus there are no test changes here.

However, it was reporting the error as part of validating the SUPER node,
while simultaneously incorrectly validating the OPTCHAIN_GETPROP or
OPTCHAIN_GETELEM that had SUPER as it's first child.

This little fix is really just to avoid a potential future bug when
someone refactors the code in a way that stops hiding the fact that
the validation for GETPROP and GETELEM were wrong.

PiperOrigin-RevId: 317210405