AstChainOptionalizer: utility to aid addition of optional chain support
Add a temporary utility to CompilerTestCase allowing one to make an existing
test class run all of its test cases in a way that checks equivalent cases
using optional chaining.

1. Wherever it is valid to do so, change all CALL, GETPROP, and GETELEM nodes
   in the input JS code to the OPTCHAIN_* version of that node.
2. Process the code.
3. Change all of the OPTCHAIN_* nodes in the result back to their non-OPTCHAIN_
   versions.

The idea is to use this utility in a working directory to identify problem
spots, fix them, and add a few targeted test cases for optional chains.

No test classes should be commited to the repository with this feature enabled.
It should be removed entirely once support for optional chaining is complete.

PiperOrigin-RevId: 317381855