Add template parameters for async generator return value and next(value)
The template parameters are not yet used to affect the prototype methods.
That requires more extensive changes and probably fixes to client code.

However, this change is sufficient to avoid complaining when client code
uses 3 template parameters, as happens when the async generator d.ts definitions
are used with tsickle to generate closure JS from TS.

PiperOrigin-RevId: 317743376