Remove "normalizing" of block scoped function declarations to avoid changing function scoping.
When this was originally written block scoped function were not part of the language spec and the behavior varied by browser.  ECMAScript 5 made them illegal in strict mode.

These days, by default, block scope functions are rewritten to conform to ES6 strict mode scoping rules when targetting ES5 and this is no longer required.

When targetting ES2015 or later, this rewriting isn't necessary and incorrectly changes the scoping of the functions.  Removing the rewriting fixes these issues.

PiperOrigin-RevId: 322899211