ExpressionDecomposerTest: allow testing arbitrary expressions
Previously the testing methods in ExpressionDecomposerTest allowed for
identification of expressions to be tested only by providing a string that
was expected to either match the callee of a CALL node or to be "yield" to
match the first found `yield` expression.

This change adds methods that instead allow matching an arbitrary expression
by supplying a string to match. The existing methods have been refactored
to be thin wrappers around the newer ones.

More cleanup should probably be done. In particular, I think I'd rather delete the
methods that just take a string and match it against CALL or YIELD nodes, because
I think the test cases would be more readable if they all used the string expression
matching I've added. However, doing that now would cause merge conflicts with
the changes I was already working on when I discovered I needed this
capability, so I'd prefer to do it later.

PiperOrigin-RevId: 323610458