Align the diagnostics in CheckMissingRequires
Make the MISSING_REQUIRE diagnostics default to warning, and the
MISSING_REQUIRE_TYPE diagnositcs default to disabled.

This is a step toward unifying behavior of this pass inside goog.provide files
and outside, and may also help with replacing the CheckMissingAndExtraRequires
pass, which doesn't have an equivalent of the MISSING_REQUIRE_TYPE diagnostics.

PiperOrigin-RevId: 325847664