Differentiate the names of the two types called `Function.prototype`.
One of these types is the implicit prototype of all functions, as well as the ".protoype" property of `(typeof Function)`. This type continues to be called `Function.prototype`.

The other is the ".prototype" property of the type `Function` ~= `function(new:?, ...?):?`. This type is now called `?.prototype`, since its owner constructor constructs `?`s.

PiperOrigin-RevId: 326565952