Clean up GlobalNamespaceTest
GlobalNamespaceTest requires an unusual setup and doesn't use CompilerTestCase.
This change makes the intentions of that setup clearer to the reader.

Specifically, I was surprised by failures in this test when I needed to change
the language input level.  The test requires that no transpilation occur, but
the setup code didn't make this intention clear.

PiperOrigin-RevId: 329995574