Fix test infrastructure that causes MaybeReachingVariableUseTest tests to silently pass.
Changing each test case's `assertMatch` to `assertNotMatch` made no difference as all unit tests in MaybeReachingVariableUseTest still passed. This is because the `assertNoMatch` function incorrectly passes an ArrayList `uses` to the `assertThat(result).doesNotContain(uses)` (that checks that the `result` does not contain `uses` ArrayList, instead of checking that `result` does not contain the inner *elements* of `uses`).

PiperOrigin-RevId: 330519158