Treat FUNCTION_INSTANCE_PROTOTYPE (aka `?.prototype`) as invalidating.
This is done by adding a `JSTypeNative` entry for FUNCTION_INSTANCE_PROTOTYPE. The name doesn't match the pattern of the other native type names, but `Function` is strange.

Without this, there are issues when ctors are assigned to a `Function` reference, because mutating the `.prototype` of that reference can't be traced back to the real ctor. This can even happen accidentally because inference sometimes defaults to `Function` for complex cases.

PiperOrigin-RevId: 330524797