Roll forward of "Stop using JSType in RemoveUnusedCode"
RemoveUnusedCode attempts to distinguish between properties set on constructors
from those set on other kinds of objects. It may be able to remove properties
set on constructors if they are never used.

Prior to this change, when it considered `X.prop = value`, it would look at the
JSType information of `X` to determine whether `X` was a constructor.
However, that is problematic for the future direction we want to take to use
less type information in optimization passes.

With this change, RemoveUnusedCode instead looks at what value was assigned to
`X`. If the value is a function or class literal, it will treat `X.prop` as if
it were set on a constructor.

This change is expected to make very little difference to the optimized output,
but there are some subtle differences.

PiperOrigin-RevId: 330747616