Roll forward of "Stop using JSType in RemoveUnusedCode"
RemoveUnusedCode attempts to distinguish between properties set on constructors
from those set on other kinds of objects. It may be able to remove properties
set on constructors if they are never used.

Prior to this change, when it considered `X.prop = value`, it would look at the
JSType information of `X` to determine whether `X` was a constructor.
However, that is problematic for the future direction we want to take to use
less type information in optimization passes.

With this change, RemoveUnusedCode instead looks at what value was assigned to
`X`. If the value is a function or class literal, it will treat `X.prop` as if
it were set on a constructor.

This change is expected to make very little difference to the optimized output,
but there are some subtle differences.

NEW:
This is the same as the change that was rolled back, but I needed to fix some
project externs to avoid incorrect code removal before rolling this forward.

This included adding a missing property to the angular-1.6 externs

https://github.com/google/closure-compiler/commit/ebb1de9849ce2bb9b0f105a9fb9dc836aafb4a8b

If this change breaks you by removing definitions of `Foo.prop`, you need
to add `Foo.prop` to your externs.

PiperOrigin-RevId: 332042509