Delete redundant information from POM files.
Since builds are now conducted with Bazel, and none of the JARs we release need any dependencies, all of the build configuration in the POM files can be deleted. All that's necessary is the project and artifact info.

This change also moves/renames the POMs since they no longer need to be in the repo root.

PiperOrigin-RevId: 333171273