Fix JSCompiler crash in InlineObjectLiterals pass when inlining object literal with language out ES2017
The InlineObjectLiterals pass is incorrectly calculating the hoist scope when a local-scoped((let/const) variable's declaration is not 'well-defined' (i.e. local variable is not assigned to the object literal as part of its declaration).

Also, this happens not just when the assignment is inside a conditional, but also on an unconditional assignments. Example:

```
{let a; if(cond) {a={x:""};console.log(a.x);}}. // assignment inside conditional - crashes
{let a; a={x:""};console.log(a.x);}  // unconditional assignment - this also crashes
{let a={x:""};console.log(a.x);}}. // unconditional assignment - does not crash as declaration is 'well-defined'
```

PiperOrigin-RevId: 333454532