Initial support for dynamic imports
Allows dynamic import statements to be recognized by the compiler and utilized. Dynamic import expressions can be used as a script or module loader even without full module rewriting. Contains only basic support for type checking in that the type of the import is a Promise resolving to the unknown type.