Make navigator extern non-nullable
navigator is always present in every supported browser (see MDN link); the absence of a non-nullability annotation appears to be an oversight. This change also matches the behavior of TypeScript.

This is a backwards-compatible change for code that reads from this property. Code that's setting up a fake window or otherwise messing with the environment could in principle break, but this should be rare.

PiperOrigin-RevId: 340759506