Fix incorrect hoisting of template string arrays.
With this change, the parent class (LateEs6ToEs3Converter.java) passes-in the insertion point to the Es6TemplateLiterals class for injecting calls to `createTemplateTagFirstArg`.

The insertion point passed-in to Es6TemplateLiterals for the first script is the node after the runtime injected def of `createTemplateTagFirstArg` in that first script. Then, for any subsequent script(s), the passed-in insertion point is the  top of that script.

PiperOrigin-RevId: 343112951